Update cmder script to check if a conemu.xml exists before trying to back it up.

+++ packageScripts/cmder.ps1
@@ -2,6 +2,9 @@ $conemuDir = "C:\tools\cmder\vendor\conemu-maximus5\"
 echo "Copying conemu config from git to $conemuDir"
 pushd $conemuDir
 $date = get-date -Format "yyyyMMdd_h-m-s"
-mv ConEmu.xml "ConEmu-$date.xml"
+if (Test-Path "$conemuDir\ConEmu.xml")
+{
+    mv ConEmu.xml "ConEmu-$date.xml"
+}
 cp "$PSScriptRoot\cmder\ConEmu.xml" .
 popd
